Shading isn't enough - More necessary
In eclipse, when the mockito-shaded project is open, it causes errors to
show elsewhere because there is no actual source code.  This prevents
eclipse resolving correctly.  Further, correcting the issue showed that
the shading was not fully completed as such it had errors on source side
of the house.  So after a number of adjustments eclipse now shows
without errors and shading occurs as needed.  No actual code changes
necessary.